@@ -1,6 +1,6 @@
 /*****************************************************************************************
  *
- * Copyright 2016 Gregory Brown. All Rights Reserved.
+ * Copyright 2016-2025 Gregory Brown. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -19,240 +19,102 @@
 
 package com.gabstudios.cmdline;
 
-import java.util.LinkedList;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
-
-import com.gabstudios.logging.LogProvider;
-import com.gabstudios.logging.LogService;
-
+import java.util.stream.Collectors;
 
 /**
- * This class is used to tokenize the values from the command line.
- * 
- * This parser is able to handle equals, commas and a value.
- * 
- * The order of priority is: equals, commas and value.
- * 
- * It will produce a List of String values.
+ * Tokenizes command definitions into {@link Token} objects based on prefixes. Supports the following prefixes: -
+ * {@code #} : Description - {@code !} : Required value (or {@code !...} for required list) - {@code ?} : Optional value
+ * (or {@code ?...} for optional list) - {@code :} : Regex validation - No prefix: Command name Splits on {@code =} and
+ * {@code ,} delimiters, prioritizing equals over commas.
  *
  * @author Gregory Brown (sysdevone)
- *
  */
-public class DefinedCommandTokenizer
-{
-    /*
-     * Constructor.
+public class DefinedCommandTokenizer {
+
+    private static final int MAX_TOKEN_LENGTH = 1000; // Configurable max length for security
+
+    /**
+     * Protected constructor to prevent instantiation.
      */
-    protected DefinedCommandTokenizer()
-    {
-        // void - do nothing.
+    protected DefinedCommandTokenizer() {
+        // Do nothing
     }
-    
-    /*
-     * Creates a token based on the first char.
-     * 
-     * A String must use one of these symbols in order for it to be recognized
-     * as that type:
-     * 
-     * # = The description of the command. There may be zero to one defined.
-     * 
-     * ! = A required value for the command name. There can be zero to many
-     * defined.
-     * 
-     * ? = An optional value for the command name. There can be zero to many
-     * defined.
-     * 
-     * : = The regex value to match on for any values that are defined. There
-     * can be zero to one defined.
-     * 
-     * 
+
+    /**
+     * Creates a {@link Token} based on the input string's prefix.
+     *
      * @param inputString
-     * @param tokenList
+     *            the string to tokenize (must not be null or empty)
+     *
+     * @return the corresponding {@link Token}
+     *
+     * @throws IllegalArgumentException
+     *             if input is invalid or exceeds max length
      */
-    private static void createToken(final String inputString,
-            final List<Token> tokenList)
-    {
-        
-        assert( inputString != null && inputString.length() > 0 ) : "The parameter 'inputString' must not be null or empty";
-        assert( tokenList != null) : "The parameter 'tokenList' must not be null";
-        //TODO - add assert for max length.
-        
-        if (inputString.startsWith("#"))
-        {
-            tokenList.add(new Token(Token.Type.DESCRIPTION, inputString
-                    .substring(1)));
+    private static Token createToken(String inputString) {
+        if (inputString == null || inputString.isEmpty()) {
+            throw new IllegalArgumentException("Input string must not be null or empty");
         }
-        else if (inputString.startsWith("!") || inputString.startsWith("?"))
-        {
-            DefinedCommandTokenizer.handleListValue(inputString, tokenList);
+        if (inputString.length() > MAX_TOKEN_LENGTH) {
+            throw new IllegalArgumentException("Input string exceeds maximum length of " + MAX_TOKEN_LENGTH);
         }
-        else if (inputString.startsWith(":"))
-        {
-            tokenList.add(new Token(Token.Type.REGEX_VALUE, inputString
-                    .substring(1)));
-        }
-        else
-        {
-            tokenList.add(new Token(Token.Type.COMMAND, inputString));
+
+        switch (inputString.charAt(0)) {
+            case '#':
+                return new Token(Token.Type.DESCRIPTION, inputString.substring(1));
+            case '!':
+                return createValueToken(inputString, true);
+            case '?':
+                return createValueToken(inputString, false);
+            case ':':
+                return new Token(Token.Type.REGEX_VALUE, inputString.substring(1));
+            default:
+                return new Token(Token.Type.COMMAND, inputString);
         }
     }
-    
-    /*
-     * Creates a token based on the first char.
-     * 
-     * A String must use one of these symbols in order for it to be recognized
-     * as that type:
-     * 
-     * ! = A required value for the command name. There can be zero to many
-     * defined.
-     * 
-     * ? = An optional value for the command name. There can be zero to many
-     * defined.
-     * 
-     * ... = A value ends with ... and is a list for the command name. There can
-     * be zero to one defined. This can be used with the ! and ? symbols
-     * 
-     * 
-     * @param inputString
-     * @param tokenList
+
+    /**
+     * Helper to create value or list tokens for required/optional.
      */
-    
-    private static void handleListValue(final String inputString,
-            final List<Token> tokenList)
-    {
-        
-        assert( inputString != null && inputString.length() > 0 ) : "The parameter 'inputString' must not be null or empty";
-        assert( tokenList != null) : "The parameter 'tokenList' must not be null";
-        //TODO - add assert for max length.
-        
-        if (inputString.startsWith("!") && inputString.endsWith("..."))
-        {
-            int length = inputString.length() - 3;
-            tokenList.add(new Token(Token.Type.REQUIRED_LIST_VALUE, inputString
-                    .substring(1, length)));
+    private static Token createValueToken(String inputString, boolean required) {
+        boolean isList = inputString.endsWith("...");
+        String value = isList ? inputString.substring(1, inputString.length() - 3) : inputString.substring(1);
+
+        Token.Type type;
+        switch (required ? 1 : 0) {
+            case 1:
+                type = isList ? Token.Type.REQUIRED_LIST_VALUE : Token.Type.REQUIRED_VALUE;
+                break;
+            case 0:
+                type = isList ? Token.Type.OPTIONAL_LIST_VALUE : Token.Type.OPTIONAL_VALUE;
+                break;
+            default:
+                throw new IllegalStateException("Unexpected value");
         }
-        else if (inputString.startsWith("!"))
-        {
-            tokenList.add(new Token(Token.Type.REQUIRED_VALUE, inputString
-                    .substring(1)));
-        }
-        else if (inputString.startsWith("?") && inputString.endsWith("..."))
-        {
-            int length = inputString.length() - 3;
-            tokenList.add(new Token(Token.Type.OPTIONAL_LIST_VALUE, inputString
-                    .substring(1, length)));
-        }
-        else if (inputString.startsWith("?"))
-        {
-            tokenList.add(new Token(Token.Type.OPTIONAL_VALUE, inputString
-                    .substring(1)));
-        }
+
+        return new Token(type, value);
     }
-    
-    /*
-     * Tokenize the values that were defined.
-     * 
+
+    /**
+     * Tokenizes the command definition arguments into a list of {@link Token}s.
+     *
      * @param args
-     *            An array of String containing one to many values.
-     * @return A List<Token> instance.  May return an empty List if input is empty or nothing is processed.
+     *            array of strings to tokenize (may be null or empty)
+     *
+     * @return a list of tokens (empty if no valid input)
      */
-    protected List<Token> tokenize(final String[] args)
-    {
-        //assert( args != null && args.length > 0 ) : "The parameter 'args' must not be null or empty";
-        //TODO - add assert for max length.
-        
-        // process spaces - taken care of by command line.
-        final LinkedList<Token> tokenList = new LinkedList<Token>();
-        
-        if (args != null && args.length > 0)
-        {
-            final LogService logService = LogProvider.getProvider()
-                    .getService();
-            
-            //
-            // Tokenize the values using the EQUALS as a delimiter and process.
-            // ---------------------------
-            for (String argString : args)
-            {
-                // equals
-                logService.logDebug(this.getClass(), "tokenize",
-                        "Before equals split: " + argString);
-                
-                //
-                // Tokenize the values using the EQUALS as a delimiter and process.
-                // ---------------------------
-                final String[] argsAfterEquals = argString.split("=");
-                
-                logService.logDebug(this.getClass(), "tokenize",
-                        "After equals split length: " + argsAfterEquals.length);
-                
-                //
-                // Process tokens after using EQUALS as a delimiter.
-                // ---------------------------
-                for (String argAfterEquals : argsAfterEquals)
-                {
-                    
-                    // commas
-                    
-                    logService.logDebug(this.getClass(), "tokenize",
-                            "Before comma split: " + argAfterEquals);
-                    
-                    //
-                    // Tokenize the values using the COMMA as a delimiter and process.
-                    // ---------------------------
-                    final String[] argsAfterCommas = argAfterEquals.split(",");
-                    
-                    logService.logDebug(this.getClass(), "tokenize",
-                            "After comma split length: "
-                                    + argsAfterCommas.length);
-                    //
-                    // Process tokens after using COMMA as a delimiter.
-                    // ---------------------------
-                    for (String argAfterComma : argsAfterCommas)
-                    {
-                        
-                        // values
-                        if (argAfterComma.length() > 0)
-                        {
-                            DefinedCommandTokenizer.createToken(argAfterComma,
-                                    tokenList);
-                        }
-                    }
-                    //
-                    // After processing args for commas, add the previous token if it is not a comma. 
-                    // ---------------------------
-                    if (argsAfterCommas.length == 0)
-                    {
-                        if (argAfterEquals.equals(","))
-                        {
-                            // ignore.
-                        }
-                        else if (argAfterEquals.length() > 0)
-                        {
-                            DefinedCommandTokenizer.createToken(argAfterEquals,
-                                    tokenList);
-                        }
-                    }
-                    
-                }
-                
-                //
-                // After processing args for equals, add the previous token if it is not a equal. 
-                // ---------------------------
-                if (argsAfterEquals.length == 0)
-                {
-                    if (argString.equals("="))
-                    {
-                        // ignore.
-                    }
-                    else if (argString.length() > 0)
-                    {
-                        tokenList.add(new Token(Token.Type.COMMAND, argString));
-                    }
-                }
-                
-            }
+    protected List<Token> tokenize(String[] args) {
+        if (args == null || args.length == 0) {
+            return new ArrayList<>();
         }
-        return (tokenList);
+
+        return Arrays.stream(args).flatMap(arg -> Arrays.stream(arg.split("="))) // Split on '=', flatten
+                .flatMap(eqPart -> Arrays.stream(eqPart.split(","))) // Split on ',', flatten
+                .filter(s -> !s.isEmpty()) // Ignore empty strings
+                .map(s -> createToken(s)) // Create tokens (lambda for explicit type inference)
+                .collect(Collectors.toList());
     }
 }
