@@ -1,33 +1,30 @@
 /*****************************************************************************************
- * 
- * Copyright 2016 Gregory Brown. All Rights Reserved.
- * 
+ *
+ * Copyright 2016-2025 Gregory Brown. All Rights Reserved.
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations under
  * the License.
- * 
- ***************************************************************************************** 
+ *
+ *****************************************************************************************
  */
 
 package com.gabstudios.cmdline;
 
 /**
  * An exception that is used if a match fails when the regex value has been defined.
- * 
- * 
+ *
  * @author Gregory Brown (sysdevone)
- * 
  */
-public class MatchException extends RuntimeException
-{
+public class MatchException extends RuntimeException {
 
     /**
      * Serialized version number.
@@ -36,12 +33,11 @@ public class MatchException extends RuntimeException
 
     /**
      * Constructor that takes a message.
-     * 
+     *
      * @param message
      *            A <code>String</code> message.
      */
-    protected MatchException(final String message)
-    {
+    protected MatchException(final String message) {
         super(message);
     }
 }
