@@ -1,6 +1,6 @@
 /*****************************************************************************************
  *
- * Copyright 2016 Gregory Brown. All Rights Reserved.
+ * Copyright 2016-2025 Gregory Brown. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -22,16 +22,12 @@ package com.gabstudios.cmdline;
 import java.util.ArrayList;
 import java.util.List;
 
-
 /**
  * An exception that is used if an action is not supported.
  *
- *
  * @author Gregory Brown (sysdevone)
- *
  */
-public class UnsupportedException extends RuntimeException
-{
+public class UnsupportedException extends RuntimeException {
 
     /**
      * Serialized version number.
@@ -41,7 +37,7 @@ public class UnsupportedException extends RuntimeException
     /**
      * Holds possible suggestions to return to the user.
      */
-    private List<String>      _suggestionList;
+    private List<String> _suggestionList;
 
     /**
      * Constructor that takes a message.
@@ -49,8 +45,7 @@ public class UnsupportedException extends RuntimeException
      * @param message
      *            A <code>String</code> message.
      */
-    protected UnsupportedException(final String message)
-    {
+    protected UnsupportedException(final String message) {
         super(message);
     }
 
@@ -60,12 +55,9 @@ public class UnsupportedException extends RuntimeException
      * @param message
      *            A <code>String</code> message.
      * @param suggestionList
-     *            A <code>List</code> of possible suggestions to return to the
-     *            user if the command was misspelled
+     *            A <code>List</code> of possible suggestions to return to the user if the command was misspelled
      */
-    public UnsupportedException(final String message,
-            final List<String> suggestionList)
-    {
+    public UnsupportedException(final String message, final List<String> suggestionList) {
         super(message);
         this._suggestionList = new ArrayList<String>(suggestionList);
     }
@@ -73,11 +65,9 @@ public class UnsupportedException extends RuntimeException
     /**
      * Gets a <code>List</code> of suggestion alternatives.
      *
-     * @return A <code>List</code> instance containing zero to many
-     *         <code>String</code>instances.
+     * @return A <code>List</code> instance containing zero to many <code>String</code>instances.
      */
-    public List<String> getSuggestionList()
-    {
+    public List<String> getSuggestionList() {
         return (this._suggestionList);
     }
 }
